+++ lib/services.ts
@@ -3,14 +3,16 @@ import { TaskStatus } from "@prisma/client";
 import { removeUndefinedProperties } from "./utils";
 
 export async function searchProjects(userId: string, searchQuery: string): Promise<Project[]> {
-  const sql = `
-    SELECT id, title, description, createdAt, updatedAt, userId 
-    FROM Project 
-    WHERE userId = '${userId}' AND title LIKE '%${searchQuery}%'
-  `;
-  
-  const results = await prismaClient.$queryRawUnsafe(sql);
-  return results as Project[];
+  const results = await prismaClient.project.findMany({
+    where: {
+      userId: userId,
+      title: {
+        contains: searchQuery,
+      },
+    },
+    orderBy: { updatedAt: "desc" },
+  });
+  return results;
 }
 
 export type Project = {